:ok_hand: introduced an Option model for the ServiceRate module to separate concerns
